serilization/deserialization and datamodel and factory seperation
